+++ ARGO/include/OptionsScreen.h
@@ -8,7 +8,7 @@ public:
 	OptionsScreen(EventManager& t_eventManager, Controller& t_controller, SDL_Renderer* t_renderer);
 	~OptionsScreen();
 
-	void update(Uint16 t_deltaTime);
+	void update(float t_deltaTime);
 	void render(SDL_Renderer* t_renderer);
 	void reset();
 	void initialise();
