+++ ARGO/src/LicenseScreen.cpp
@@ -11,7 +11,7 @@ LicenseScreen::~LicenseScreen()
 {
 }
 
-void LicenseScreen::update(Uint16 t_deltaTime)
+void LicenseScreen::update(float t_deltaTime)
 {
 	m_inputSystem.update(m_inputEntity);
 	m_commandSystem.update(m_inputEntity, m_eventManager);
